Create Photoshop.md
# Photoshop

Photoshop has a very limited tool set in 32 bit mode, so if you're wanting to edit scene-linear EXR file, a better option is Affinity Photo, which is free. 
For editing 16-bit log or 8-bit files in display-referred sRGB images there is a free [OCIO plugin for Photoshop](http://fnordware.blogspot.com/2017/02/opencolorio-for-photoshop.html).